GP plus added!!!! 🤞

+++ backend/services/unitServices.ts
@@ -15,7 +15,8 @@ import moment from "moment";
 import { nanoid } from "nanoid";
 import { getLiveUnits } from "../../shared/fns";
 import { UNITS_URL_PATH } from "../../shared/constants";
-import dbWebApps, { TWebAppForUI } from "../models/WebApp";
+import { TWebAppForUI } from "../models/WebApp";
+import dbWebApps from "../models/WebApp";
 
 const insertUnit = async (unit: INewUnitSchema) => {
   try {
@@ -45,7 +46,7 @@ const insertUnit = async (unit: INewUnitSchema) => {
   }
 };
 
-const deleteUnit = async (_id?: string, queryPair?: [string, unknown]) => {
+const deleteUnit = async (_id?: unknown, queryPair?: [string, unknown]) => {
   try {
     console.log(
       `Attempting to delete unit with id ${_id} and queryPair ${JSON.stringify(
@@ -66,13 +67,19 @@ const deleteUnit = async (_id?: string, queryPair?: [string, unknown]) => {
       };
     }
 
+    if(_id && (typeof _id !== 'string')){
+      throw new Error(
+        "`_id` must be a string."
+      );
+    }
+
     let deletionResult: DeleteResult;
 
     if (queryPair && queryPair.length > 0) {
       const [key, val] = queryPair;
       deletionResult = await Unit.deleteOne({ [key]: val });
     } else {
-      deletionResult = await Unit.deleteOne({ _id });
+      deletionResult = await Unit.deleteOne({ _id: { $eq: _id } });
     }
 
     if (deletionResult.deletedCount === 0) {
@@ -196,8 +203,6 @@ const updateUnit = async (
   customUpdate?: TCustomUpdate
 ) => {
   try {
-    // an example of lesson being updated:
-    // section.participants[0].name = "John Doe"
     if (!Unit) {
       throw new Error(
         "Failed to connect to the database. `Units` collections does not exist."
@@ -207,7 +212,7 @@ const updateUnit = async (
     if (customUpdate) {
       console.log("Making a custom update.");
 
-      const result = await Unit.updateMany(filterObj, customUpdate).lean();
+      const result = await Unit.updateMany(filterObj, customUpdate);
 
       if (result.matchedCount === 0) {
         return { errMsg: "No matching units were found in the database." };
@@ -220,7 +225,7 @@ const updateUnit = async (
 
     const { modifiedCount, matchedCount } = await Unit.updateMany(filterObj, {
       $set: updatedProps,
-    }).lean();
+    });
 
     if (matchedCount === 0) {
       return { errMsg: "No matching units were found in the database." };
@@ -527,8 +532,12 @@ const getIsUnitNew = (releaseDate: Date, now: number) => {
   return isNew;
 };
 
-const filterInShowableUnits = (units: INewUnitSchema[], nowMs: number) => {
+const filterInShowableUnits = (units: INewUnitSchema[], nowMs: number, willGetUnitMetaData = true) => {
   const liveUnits = getLiveUnits(units).filter((unit) => unit?.ReleaseDate);
+
+  if(!willGetUnitMetaData){
+    return liveUnits as ILiveUnit[];
+  }
 
   return liveUnits.map((unit) => {
     const individualLessonsNum =
